Simplify client router to be singleton
