+++ lib/link.js
@@ -1,6 +1,6 @@
 import React, { Component, Children } from 'react'
 import PropTypes from 'prop-types'
-import Router from './router'
+import Router from '../client/router'
 import { warn, execOnce } from './utils'
 import { formatPath } from './url'
