+++ lib/page-loader.js
@@ -1,8 +1,8 @@
 /* global window, document */
-import EventEmitter from './EventEmitter'
+import mitt from 'mitt'
 
 const pageCache = {}
-const pageRegisterEvents = new EventEmitter()
+const pageRegisterEvents = mitt()
 
 export function waitForPage (route) {
   route = normalizeRoute(route)
