+++ package.json
@@ -45,6 +45,7 @@
   "dependencies": {
     "del": "3.0.0",
     "htmlescape": "1.1.1",
+    "mitt": "1.1.3",
     "minimist": "1.2.0",
     "prop-types": "15.6.2",
     "querystring": "0.2.0",
